Improve name lookups
Adds --mc.name-source flag that allows to change behavior when looking up for player name. Supported values are "mojang" and "bukkit". If set to "mojang", it will make request to Mojang API for each player. If set to "bukkit", name will be fetched from NBT tag "lastKnownName" that added by Bukkit. If "offline" or any other value is set, it will use UUID as name.